+++ src/platforms/aworkslp/serial_helper.c
@@ -29,7 +29,6 @@
 
 #include "tkc/mem.h"
 #include "tkc/wstr.h"
-#include "tkc/thread.h"
 #include "tkc/socket_pair.h"
 #include "tkc/socket_helper.h"
 #include "streams/serial/serial_helper.h"
@@ -42,16 +41,12 @@
 #include "aw_unistd.h"
 
 #include "aw_serial.h"
-#include "tkc/ring_buffer.h"
 #include "tkc/darray.h"
 
 static darray_t* s_darray = NULL;
 
 typedef struct {
-  tk_thread_t* thread;
   serial_handle_t handle;
-  bool_t broken;
-  ring_buffer_t* rb;
   bool_t is_usb2uart;
 } serial_dev_prv_t;
 
@@ -66,34 +61,10 @@ static int da_serial_cmp(const void* a, const void* b) {
 static ret_t da_serial_destroy(void* data) {
   serial_dev_prv_t* p_this = data;
 
-  if (p_this->thread) {
-    p_this->broken = TRUE;
-    ENSURE(tk_thread_join(p_this->thread) == RET_OK);
-    tk_thread_destroy(p_this->thread);
-  }
-
-  ring_buffer_destroy(p_this->rb);
   TKMEM_FREE(p_this);
   return RET_OK;
 }
 
-static void* thread_serial_recv(void* args) {
-  serial_dev_prv_t* p_this = args;
-  int32_t cnt = 0;
-  int fd = serial_handle_get_fd(p_this->handle);
-  uint8_t buff[64];
-  ENSURE(p_this && fd >= 0);
-
-  while (!p_this->broken) {
-    cnt = aw_read(fd, buff, sizeof(buff));
-    if (cnt > 0) {
-      ring_buffer_write_len(p_this->rb, buff, cnt);
-    } else {
-      sleep_ms(10);
-    }
-  }
-  return NULL;
-}
 
 int serial_handle_get_fd(serial_handle_t handle) {
   return handle->dev;
@@ -118,8 +89,6 @@ serial_handle_t serial_open(const char* port) {
 
   memset(p_this, 0, sizeof(serial_dev_prv_t));
   p_this->handle = handle;
-  p_this->rb = ring_buffer_create(256, 4096);
-  ENSURE(p_this->rb);
 
   if (strstr(port, "USB")) {
     p_this->is_usb2uart = TRUE;
@@ -141,27 +110,31 @@ ret_t serial_iflush(serial_handle_t handle) {
 
 ret_t serial_oflush(serial_handle_t handle) {
   int fd = serial_handle_get_fd(handle);
-  serial_dev_prv_t* p_this = darray_find(s_darray, handle);
-  return_value_if_fail(p_this, RET_FAIL);
-  ring_buffer_reset(p_this->rb);
   return aw_ioctl(fd, AW_FIOWFLUSH, NULL) == 0 ? RET_OK : RET_FAIL;
 }
 
 ret_t serial_wait_for_data(serial_handle_t handle, uint32_t timeout_ms) {
-  serial_dev_prv_t* p_this = darray_find(s_darray, handle);
-  return_value_if_fail(p_this, RET_FAIL);
-  if (ring_buffer_size(p_this->rb)) {
-    return RET_OK;
-  }
-  sleep_ms(timeout_ms);
-  return RET_TIMEOUT;
+
+    int len = 0;
+    int fd = serial_handle_get_fd(handle);
+    while (timeout_ms) {
+      aw_ioctl(fd, AW_FIONREAD, &len);
+      if (len > 0) {
+        return RET_OK;
+      }
+      sleep_ms(10);
+      timeout_ms -= tk_min(timeout_ms, 10);
+    }
+    return RET_TIMEOUT;
+
 }
 
 int32_t serial_read(serial_handle_t handle, uint8_t* buff, uint32_t max_size) {
   int fd = serial_handle_get_fd(handle);
-  serial_dev_prv_t* p_this = darray_find(s_darray, handle);
+  int cnt = 0;
   return_value_if_fail(fd >= 0, RET_FAIL);
-  return ring_buffer_read(p_this->rb, buff, max_size);
+  cnt = aw_read(fd, buff, max_size);
+  return cnt;
 }
 
 int32_t serial_write(serial_handle_t handle, const uint8_t* buff, uint32_t max_size) {
@@ -174,7 +147,6 @@ int32_t serial_write(serial_handle_t handle, const uint8_t* buff, uint32_t max_s
 
 int serial_close(serial_handle_t handle) {
   int fd = serial_handle_get_fd(handle);
-  serial_dev_prv_t* p_this = darray_find(s_darray, handle);
   return_value_if_fail(aw_close(fd) == 0, RET_FAIL);
   darray_remove(s_darray, handle);
   return RET_OK;
@@ -220,13 +192,6 @@ ret_t serial_config(serial_handle_t handle, uint32_t baudrate, bytesize_t bytesi
     /* 配置串口超时 */
     ret = aw_serial_timeout_set(fd, &timeout);
     return_value_if_fail(ret == RET_OK, RET_FAIL);
-  }
-
-  if (!p_this->thread) {
-    p_this->thread = tk_thread_create(thread_serial_recv, p_this);
-    //  tk_thread_set_priority(p_this->thread, 6);
-    tk_thread_set_stack_size(p_this->thread, 2048);
-    ENSURE(tk_thread_start(p_this->thread) == RET_OK);
   }
 
   return RET_OK;